renamed manager.py to config_manager.py and added more docstrings
